get all trips from database

+++ Easyride/src/main/java/com/Easyride/Trip/Trip.java
@@ -2,7 +2,6 @@ package com.Easyride.Trip;
 
 import com.Easyride.Driver.Driver;
 import com.Easyride.Rider.Rider;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.Data;
 
 import javax.persistence.*;
@@ -14,10 +13,10 @@ public class Trip {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     int id;
-    private long pickUpLat;
-    private long pickUpLng;
-    private long dropLat;
-    private long dropLng;
+    private float pickUpLat;
+    private float pickUpLng;
+    private float dropLat;
+    private float dropLng;
     private String time;
     private String day;
